refactor: Moved the warning for settings to the domain list area
